requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
backports.zoneinfo==0.2.1; python_version < '3.9'
dj-database-url==0.5.0
django-filter==21.1
django-heroku==0.3.1
django==4.0
djangorestframework==3.13.1
faker==10.0.0
gunicorn==20.1.0
importlib-metadata==4.10.0; python_version < '3.10'
markdown==3.3.6
psycopg2==2.9.2; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
text-unidecode==1.3
whitenoise==5.3.0
zipp==3.6.0; python_version >= '3.6'
